added Seller Dashboard + resolve config after pull all work to my computer
