+++ next.config.js
@@ -1,0 +1,9 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  // ✅ use config options as before to change .ts to .js
+  experimental: {
+    optimizeCss: false,
+  },
+};
+
+module.exports = nextConfig;
