--- next.config.ts
@@ -1,10 +1,0 @@
-import type { NextConfig } from "next";
-
-const nextConfig: NextConfig = {
-  /* config options here */
-  experimental: {
-    optimizeCss: false
-  }
-};
-
-export default nextConfig;
