+++ package.json
@@ -13,30 +13,30 @@
   },
   "dependencies": {
     "@prisma/client": "^6.12.0",
-    "autoprefixer": "^10.4.21",
-    "bcrypt": "^6.0.0",             
+    "bcrypt": "^6.0.0",
+    "bcryptjs": "^3.0.2",
     "bootstrap": "^5.3.7",
     "jsonwebtoken": "^9.0.2",
-    "next": "15.4.1",
+    "next": "14.1.0",
     "next-auth": "^4.24.11",
-    "react": "19.1.0",
+    "react": "18.2.0",
     "react-bootstrap": "^2.10.10",
-    "react-dom": "19.1.0"
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
     "@types/bcrypt": "^6.0.0",
     "@types/jsonwebtoken": "^9.0.10",
     "@types/node": "^20",
-    "@types/react": "^19.1.8",
-    "@types/react-dom": "^19.1.6",
+    "@types/react": "^18.2.38",
+    "@types/react-dom": "^18.2.18",
     "autoprefixer": "^10.4.21",
-    "eslint": "^9",
-    "eslint-config-next": "15.4.1",
+    "eslint": "^8.56.0",
+    "eslint-config-next": "14.1.0",
     "postcss": "^8.5.6",
     "prisma": "^6.12.0",
-    "tailwindcss": "^4.1.11",
+    "tailwindcss": "^3.4.1",
     "ts-node": "^10.9.2",
-    "typescript": "5.8.3"
+    "typescript": "5.2.2"
   }
 }
