+++ src/app/cart/page.tsx
@@ -1,0 +1,3 @@
+export default function CartPage() {
+  return <h1>Cart Page</h1>;
+}
