+++ src/app/globals.css
@@ -1,15 +1,17 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-@import "tailwindcss";
 
+html, body {
+  overflow-x: hidden;
+}
 
 .hero-bg {
   background-image: url('/images/hero.jpg');
 }
 
-@layer utilities{
-  h1{
-     font-size: larger;
+@layer utilities {
+  h1 {
+    font-size: larger;
   }
 }
