+++ src/app/lib/auth.ts
@@ -1,0 +1,63 @@
+// src/lib/auth.ts
+import { AuthOptions } from "next-auth";
+import CredentialsProvider from "next-auth/providers/credentials";
+import prisma from "@/app/lib/prisma";
+import { compare } from "bcryptjs";
+import type { Session, User } from "next-auth";
+import type { JWT } from "next-auth/jwt";
+
+// ✅ add CustomUser interface outside function
+interface CustomUser extends User {
+  role: string;
+}
+
+export const authOptions: AuthOptions = {
+  providers: [
+    CredentialsProvider({
+      name: "Credentials",
+      credentials: {
+        email: { label: "Email", type: "text" },
+        password: { label: "Password", type: "password" },
+      },
+      async authorize(credentials) {
+        const user = await prisma.user.findUnique({
+          where: { email: credentials?.email },
+        });
+
+        if (!user || !user.password) throw new Error("Invalid email or password");
+
+        const isValid = await compare(credentials!.password, user.password);
+        if (!isValid) throw new Error("Invalid email or password");
+
+        return {
+          id: user.id,
+          email: user.email,
+          name: user.name,
+          role: user.role,
+        };
+      },
+    }),
+  ],
+  session: {
+    strategy: "jwt",
+  },
+  pages: {
+    signIn: "/login",
+  },
+  callbacks: {
+    async jwt({ token, user }: { token: JWT; user?: CustomUser }) {
+      if (user) {
+        token.id = user.id;
+        token.role = user.role;
+      }
+      return token;
+    },
+    async session({ session, token }: { session: Session; token: JWT }) {
+      if (token && session.user) {
+        session.user.id = token.id as string;
+        session.user.role = token.role as string;
+      }
+      return session;
+    },
+  },
+};
