+++ tsconfig.json
@@ -13,6 +13,8 @@
     "isolatedModules": true,
     "jsx": "preserve",
     "incremental": true,
+    "strictFunctionTypes": false,   // ✅ added
+    "noImplicitReturns": false,     // ✅ added
     "plugins": [
       {
         "name": "next"
@@ -22,8 +24,13 @@
       "@/*": ["./src/*"]
     }
   },
-  
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/app/AboutUs/accordion.jsx"],
+  "include": [
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    ".next/types/**/*.ts",
+    "src/app/AboutUs/accordion.jsx"
+, "next.config.js"  ],
   "exclude": ["node_modules"],
   "types": ["next-auth"]
 }
